refactor: moved all packages in pkg directory to internal
